+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 * Fix: Markata tui will remain running even when the runner fails 0.6.0.dev13 #107
 * Fix: Pinned to `textual<0.2.0` due to breaking changes 0.6.0.dev13 #107
 * Fix: snyk remove seuptools from requirements 0.6.0.dev14 #130
+* Fix: Markata.filter was missing post and m 0.6.0.dev14 #133
 
 ## 0.5.2